hyfed-compensator/requirements.txt
Django==3.2
djangorestframework==3.12.2
numpy==1.19.5
requests==2.25.1
